refactor(#245): entities/todo
