+++ src/features/todo/constants/todoColumnSort.ts
@@ -1,4 +1,4 @@
-import type { TodoListSort } from "@/entities/todo/types/types";
+import type { TodoListSort } from "@/entities/todo";
 
 const TODO_LIST_SORT_BY_LABEL_CONST = {
   "마감일 순": "DUE_DATE",
