+++ src/widgets/team/TodoSection/TodoList.tsx
@@ -2,7 +2,7 @@
 
 import { useMemo, useState } from "react";
 
-import { todoQueries } from "@/entities/todo/query/todo.queryKey";
+import { todoQueryOptions } from "@/entities/todo";
 import {
   TODO_COLUMN_DEFAULT_SORT_LABEL,
   TODO_COLUMN_SORT_LABEL_ORDER,
@@ -27,7 +27,7 @@ export const TodoList = ({ goalId, keyword, isMyTodo }: TodoListProps) => {
 
   const infiniteQueryOptions = useMemo(
     () =>
-      todoQueries.todoListInfinite(goalId, {
+      todoQueryOptions.todoListInfinite(goalId, {
         keyword,
         isMyTodo,
         sort,
